user types, connected carts to individual users, added password confirmation/sign-in and out, prevented regular users from viewing admin pages, only show items in a specific user's cart, blocked non-logged in users from adding to cart or accessing cart page, maybe some more stuff

+++ src/main/java/coffeeShopWebApp/CoffeeShop/CoffeeController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import coffeeShopWebApp.CoffeeShop.dao.ItemsDao;
 import coffeeShopWebApp.CoffeeShop.dao.UsersDao;
@@ -36,14 +37,45 @@ public class CoffeeController {
 
 	// admin menu page
 	@RequestMapping("/admin")
-	public ModelAndView adminPage() {
+	public ModelAndView adminPage(@SessionAttribute(name="user", required=false) User user) {
+		if (user!=null && user.getUsertype().equals("admin")) {
 		return new ModelAndView("admin-page");
+		}else {
+			return new ModelAndView("access-error");
+		}
 	}
 
+	//login page
+	@RequestMapping("/login")
+	public ModelAndView loginPage() {
+		return new ModelAndView("login");
+	}
+	@PostMapping("/login")
+	public ModelAndView submitLogin(@RequestParam("username") String username,
+									@RequestParam("password") String password,
+										HttpSession session, RedirectAttributes redir) {
+		User user = usersDao.findByUsername(username);
+		if(user==null || !user.getPassword().equals(password)) {
+			return new ModelAndView("login","message","incorrect username or password");
+			}
+		session.setAttribute("user", user);
+		
+		redir.addFlashAttribute("message", "Welcome, thx for logging in");
+		return new ModelAndView("redirect:/");
+	}
+	
+	@RequestMapping("/logout")
+	public ModelAndView logout(HttpSession session, RedirectAttributes redir) {
+		session.invalidate();
+		redir.addFlashAttribute("Logged Out");
+		return new ModelAndView("redirect:/");
+	}
+	
 	// view cart page
 	@RequestMapping("/cart")
-	public ModelAndView cartPage() {
-		List<CartItem> myCartItemList = cartItemsDao.findAll();
+	public ModelAndView cartPage(@SessionAttribute(name="user", required=false) User user) {
+		if(user!=null) {
+		List<CartItem> myCartItemList = cartItemsDao.findByUserId(user.getId());
 		Double priceTotal = 0.00;
 		for (int i = 0; i < myCartItemList.size(); i++) {
 			priceTotal += myCartItemList.get(i).getPriceTotal();
@@ -53,13 +85,18 @@ public class CoffeeController {
 		mav.addObject("priceTotal", priceTotal);
 //		return new ModelAndView("show-cart","cartItems",myCartItemList);
 		return mav;
+		}else {return new ModelAndView("login-request");}
 	}
 
 	// list of items you can update, delete, or add
 	@RequestMapping("/admin/items")
-	public ModelAndView itemAdmin() {
-		List<Item> myItemList = itemsDao.findAll();
-		return new ModelAndView("item-list-admin", "items", myItemList);
+	public ModelAndView itemAdmin(@SessionAttribute(name="user", required=false) User user) {
+			if (user!=null && user.getUsertype().equals("admin")) {
+				List<Item> myItemList = itemsDao.findAll();
+				return new ModelAndView("item-list-admin", "items", myItemList);
+			}else {
+				return new ModelAndView("access-error");
+			}
 	}
 
 	// page for users to add items to cart
@@ -68,16 +105,18 @@ public class CoffeeController {
 		List<Item> myItemList = itemsDao.findAll();
 		return new ModelAndView("item-list", "items", myItemList);
 	}
-
 	@PostMapping("/item-list")
-	public ModelAndView addToCart(Integer quantity, Long item_id) {
+	public ModelAndView addToCart(@SessionAttribute(name="user", required=false) User user, Integer quantity, Long item_id, Long user_id) {
+		if (user!=null) {
+			
+		
 		Item currItem = itemsDao.findById(item_id);
 		if (quantity > currItem.getQuantity()) {
 			return new ModelAndView("overquantity-error-page"); // this checks if quantity added is over stock, but not
 																// if combined quantity is over stock
 		}
 
-		List<CartItem> list = cartItemsDao.findAll();
+		List<CartItem> list = cartItemsDao.findByUserId(user_id);
 		List<Long> idList = new ArrayList<>();
 		if (list.size() != 0) {
 			for (int i = 0; i < list.size(); i++) { // this gets a list of item IDs from the cartItems in the cart to
@@ -89,39 +128,35 @@ public class CoffeeController {
 		}
 		System.out.println(list);
 		if (idList.contains(item_id)) {
-			CartItem oldcartitem = cartItemsDao.findByItemID(item_id);
-			if (oldcartitem.getQuantity() + quantity <= itemsDao.findById(item_id).getQuantity() && quantity > 0) {// this
-																													// checks
-																													// available
-																													// quantity
-																													// and
-																													// send
-																													// error
-																													// message
-																													// if
-																													// there
-																													// aren't
-																													// enough
-																													// available
+			CartItem oldcartitem = cartItemsDao.findByUserAndItem(user_id, item_id);
+			
+			// this checks available quantity and send error message if there aren't enough available
+			if (oldcartitem.getQuantity() + quantity <= itemsDao.findById(item_id).getQuantity() && quantity > 0) {
 				oldcartitem.setQuantity(quantity + oldcartitem.getQuantity());
 				cartItemsDao.update(oldcartitem);
 			} else {
 				return new ModelAndView("overquantity-error-page");
 			}
 		} else {
-			cartItemsDao.create(quantity, item_id);
+			cartItemsDao.create(quantity, item_id, user_id);
 		}
 		return new ModelAndView("redirect:/item-list");
+		}else {return new ModelAndView("login-request");}
 	}
 
 	// admin page for updating items
 	@RequestMapping("/item/update")
-	public ModelAndView showUpdateForm(@RequestParam("id") Long id) {
-		ModelAndView mav = new ModelAndView("item-form");
-		mav.addObject("item", itemsDao.findById(id));
-		mav.addObject("title", "Edit Items");
-		return mav;
-	}
+	public ModelAndView showUpdateForm(@RequestParam("id") Long id,@SessionAttribute(name="user", required=false) User user) {
+			if (user!=null && user.getUsertype().equals("admin")) {
+				ModelAndView mav = new ModelAndView("item-form");
+				mav.addObject("item", itemsDao.findById(id));
+				mav.addObject("title", "Edit Items");
+				return mav;
+			}else {
+				return new ModelAndView("access-error");
+			}
+		}
+	
 
 	@PostMapping("/item/update")
 	public ModelAndView submitUpdate(Item item) {
@@ -130,17 +165,17 @@ public class CoffeeController {
 	}
 
 	@RequestMapping("/cart/quantity-update")
-	public ModelAndView showQuantityUpdateForm(@RequestParam("id") Long id) {
+	public ModelAndView showQuantityUpdateForm(@RequestParam("id") Long id, @SessionAttribute(name="user", required=false) User user) {
 		ModelAndView mav = new ModelAndView("update-cart-quantity-form");
-		mav.addObject("cartItem", cartItemsDao.findByID(id));
+		mav.addObject("cartItem", cartItemsDao.findByUserAndItem(user.getId(), cartItemsDao.findByID(id).getItem().getId()));
 		return mav;
 	}
 
 	@PostMapping("/cart/quantity-update")
-	public ModelAndView submitQuantityUpdate(Integer quantity, Long id) {
+	public ModelAndView submitQuantityUpdate(Integer quantity, Long id, @SessionAttribute(name="user", required=false) User user) {
 		if (quantity <= cartItemsDao.findByID(id).getItem().getQuantity() && quantity > 0) { // checks to make sure the
 																								// new quantity is valid
-			CartItem tempCartItem = cartItemsDao.findByID(id);
+			CartItem tempCartItem = cartItemsDao.findByUserAndItem(user.getId(), cartItemsDao.findByID(id).getItem().getId());
 			tempCartItem.setQuantity(quantity);
 			cartItemsDao.update(tempCartItem);
 			return new ModelAndView("redirect:/cart");
@@ -154,6 +189,25 @@ public class CoffeeController {
 	public ModelAndView userReg() {
 		return new ModelAndView("user-form");
 	}
+	// creates the session and sends to registration confirmation page
+		@PostMapping("/user-registration")
+		public ModelAndView userSubmit(User user, @RequestParam("passwordConf") String passwordConf,HttpSession session, RedirectAttributes redir) {
+			User userCheck = usersDao.findByUsername(user.getUsername());
+			if (userCheck != null) {
+				return new ModelAndView("user-form", "message", "A user with that username already exists.");
+			}
+			
+			if (!passwordConf.equals(user.getPassword())) {
+				return new ModelAndView("user-form", "message", "Passwords do not match.");
+			}
+			
+			usersDao.create(user);
+			session.setAttribute("user", user);
+			String firstname = user.getFirstname();
+			ModelAndView mav = new ModelAndView("user-confirmation");
+			mav.addObject("firstname", firstname);
+			return mav;
+		}
 
 	// page letting user know they have been registered -- this link may be obsolete
 	@RequestMapping("/user-confirmation")
@@ -161,22 +215,18 @@ public class CoffeeController {
 		return new ModelAndView("user-confirmation");
 	}
 
-	// creates the session and sends to registration confirmation page
-	@RequestMapping("/submit-user-registration")
-	public ModelAndView userSubmit(User user, HttpSession session) {
-		usersDao.create(user);
-		session.setAttribute("profile", user);
-		String firstname = user.getFirstname();
-		ModelAndView mav = new ModelAndView("user-confirmation");
-		mav.addObject("firstname", firstname);
-		return mav;
-	}
+	
 
 	// add item page
 	@RequestMapping("/addItem")
-	public ModelAndView addItem() {
-		return new ModelAndView("item-form", "title", "Add an Item");
+	public ModelAndView addItem(@SessionAttribute(name="user", required=false) User user) {
+			if (user!=null && user.getUsertype().equals("admin")) {
+				return new ModelAndView("item-form", "title", "Add an Item");
+			}else {
+				return new ModelAndView("access-error");
+			}
 	}
+	
 
 	@PostMapping("/addItem")
 	public ModelAndView submitNewItem(Item item) {
@@ -197,9 +247,13 @@ public class CoffeeController {
 
 	// thing that actually deletes the item
 	@RequestMapping("/item/delete")
-	public ModelAndView deleteItem(@RequestParam("id") Long id) {
-		itemsDao.delete(id);
-		return new ModelAndView("redirect:/admin/items");
+	public ModelAndView deleteItem(@RequestParam("id") Long id,@SessionAttribute(name="user", required=false) User user) {
+			if (user!=null && user.getUsertype().equals("admin")) {
+				itemsDao.delete(id);
+				return new ModelAndView("redirect:/admin/items");
+			}else {
+				return new ModelAndView("access-error");
+			}
 	}
 
 	// delete confirmation page
@@ -216,7 +270,7 @@ public class CoffeeController {
 	}
 
 	@RequestMapping("/checkout")
-	public ModelAndView checkout() {
+	public ModelAndView checkout(@SessionAttribute(name="user", required=false) User user) {
 		Double priceTotal = 0.00;
 		List<CartItem> myCartItemList = cartItemsDao.findAll();
 		for (int i = 0; i < myCartItemList.size(); i++) {
@@ -224,7 +278,7 @@ public class CoffeeController {
 		}
 		// should probably have something that updates item stock, adds order to another
 		// database, and comfirms cc info
-		cartItemsDao.deleteAll();
+		cartItemsDao.deleteCart(user.getId());
 		return new ModelAndView("order-confirmation", "priceTotal", priceTotal);
 
 	}
